+++ src/test/java/com/dss/realworld/user/api/UserControllerTest.java
@@ -1,6 +1,9 @@
 package com.dss.realworld.user.api;
 
+import com.dss.realworld.user.domain.repository.UserRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -24,6 +27,24 @@ class UserControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @BeforeEach
+    void setUp() {
+        clearTable();
+    }
+
+    @AfterEach
+    void tearDown() {
+        clearTable();
+    }
+
+    private void clearTable() {
+        userRepository.deleteAll();
+        userRepository.resetAutoIncrement();
+    }
 
     @Test
     void Should_Success_When_AddUserDtoIsNotNull() throws Exception {
