+++ src/test/java/com/dss/realworld/user/domain/UserTest.java
@@ -2,7 +2,8 @@ package com.dss.realworld.user.domain;
 
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 public class UserTest {
